# Enable fast softmax mode in FusedSDPA