--- a/examples/text-generation/run_generation.py
+++ b/examples/text-generation/run_generation.py
@@ -238,6 +238,11 @@ def setup_parser(parser):
         help="Whether to enable Habana Flash Attention in causal mode on first token generation.",
     )
     parser.add_argument(
+        "--flash_attention_fast_softmax",
+        action="store_true",
+        help="Whether to enable Habana Flash Attention in fast softmax mode.",
+    )
+    parser.add_argument(
         "--book_source",
         action="store_true",
         help="Whether to use project Guttenberg books data as input. Usefull for testing large sequence lenghts.",