@wszczurekhabana 2024-06-11
upstreamed in: https://github.com/huggingface/optimum-habana/pull/972